agents/planner/requirements.txt
# Core production dependencies
fastapi==0.104.1
uvicorn==0.24.0
pika==1.3.2
google-generativeai==0.8.3
python-dotenv==1.0.1

# Context enrichment dependencies
httpx>=0.25.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
PyGithub>=1.59.0
tavily-python>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=1.0.0
